requirements.txt
async-generator==1.10
attrs==22.2.0
boto3==1.26.99
botocore==1.29.99
charset-normalizer==3.1.0
certifi==2022.12.7
exceptiongroup==1.1.1
fake-useragent==1.1.3
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
outcome==1.2.0
pillow==9.4.0
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
wsproto==1.2.0
